Pagination and filtering for Payments.
Adds custom MonotonicCursorPagination class, which implements
a cursor-like pagination over the payments feed. The DRF's built-in
CursorPagination is not exactly a good fit here as it doesn't has
"this" page links and doesn't return links to the "previous" page
at the beginning of the dataset - so there is no easy way to check
specifically for new items.

The implementation should be solid but the tests are still missing.