API versioning with URLPathVersioning
I was experimenting with drf-openapi package (not included in this
commit) and added some API versioning as it requests.

It is useless for the current state of the demo project as there
is only a single API version (and it most likely won't evolve as
the fate of demo projects is to be abandoned), but I decided to
keep the changes just in case.

The downside of this is that CoreAPI renderer will start saying
"/{version}/" rather than "/v1/" in URLs (which is most likely
undesirable as the different API versions better be documented
separately) and I haven't yet researched how to deal with this
properly.