Improve coverage for MonotonicCursorPagination
Some code paths weren't covered by the test suite. Now they are.

I've excluded `get_schema_fields` from the coverage for now,
since testing schemas is a relatively complicated matter that
I feel is not necessary for the scope of this demo project.

A review that checks for pragmas and adds comments justifying their
existence (or removes them) should be a good idea.