Remove `tuid` from tests
I've initially thought of some tests starting multiple threads
and trying to check things concurrently, but this idea was never
implemented. An artifact of this is random UUID that was supposed
to separate tests if multiple instances of the same is running
concurrently. This is not required in the current setup, as Django
runs every test case in a transaction.

So, I'm simplifying things here.