Docstring and type annotation improvements
I've sort of neglected annotating most of the code, since it's
basically a bunch of classes conformant to Django and DRF specs
and I don't think it makes sense to copy excerpts from their docs.

However, some places are worth properly documenting and this
commit tries to improve the situation.

I'm using Sphinx-like syntax for parameter and return value
descriptions and PEP 484 type annotations. However, this is not
a library of reusable code (although it can be made so), so
I haven't set up Sphinx or checked the documentation.